added member request updates to refactor.
